Use runWhenSmart to get problem detection to work!
This requires the introduction of `ProjectCommand` that MUST
be extended for any commands that rely on a Project, since we
need that project to get the DumbService which will let us know
when we're smart. Seems like a fair trade, and I was considering
it anyway for more convenience---we can later add a `ProjectFileCommand`
to correspond with vim's `intellivim#NewCommand` to include the
Project and the current file.