First experiments with JUnit test rendering in vim
NB: This commit includes some pretty huge refactoring that should've
been in a separate commit, but I accidentally hit the "amend" mapping
before realizing what I was doing.

Anyway: Gson no longer directly handles the heavy lifting for injection;
it does the basic unpackaging into a Pojo, and returns the ICommand
wrapped in another object that stores the relevant information for
performing injection if necessary. This means that Project is now
inserted via an Injector, and that, in lieu of a proper dependency
graph, we have a "priority" system for Injectors. It's slightly less
efficient, but at our expected scale it should be negligible.

Furthermore, for ICommands that don't need *any* injection, we can
execute them immediately on the IO handler thread instead of waiting
for the UI dispatch thread. This was needed so we could quickly
fetch the tree of test nodes without worrying about a commandline
length limit (IE: IntelliJ PUSHes to Vim that it should PULL the
full node tree, aka send-to-sync).

One issue that should be addressed is *scrolling* the test output
buffer while the test is running. Currently we have to scroll
to the end and back to the start to get it to re-draw successfully
(redraw! didn't do it). Perhaps we can scroll to end, then scroll
back to wherever the cursor was previously?