Add Profiler; enable proper use of cached Projects
Profiler is a simple tool inspired by the MarkerLog (or whatever
it was called) from Volley. The API is a bit whack because it takes
an Object for "context," but this means you can grab the relevant
Profiler as long as you have access to the context; you don't have
to muck up your APIs by passing along the Profiler, which means
we can easily strip it out if it's ever is no longer useful.

Profiler confirmed that the biggest bottleneck was loading the Project
each time (which we did because reusing them previously caused weird
inconsistency errors) so I dedicated some thought into how to make it work.
It turns out, the PsiFile returned by the PsiManager was just stale,
so I figured out how to get it to refresh itself when we fetch it
via ProjectUtil.getPsiFile(), and refactored all existing code to
just @Inject the PsiFile so this version is always called.

Completion calls and file updates are now almost instantaneous!
Went from ~750ms for a completion to ~60ms. (The ensureProject
was "only" ~350ms, so I'm guessing keeping the project open lets
it cache other stuff useful for analysis, as well). This potentially
makes IntelliVim actually usable, now!

We do get a bunch of junk on stderr, but everything seems to work
anyway, and I'm hopeful that we can fix that nonsense in the future