Add initial support on IJ side for handling ambiguous imports
Refs #3

The basic idea now is that any QuickFix object can include
a "choices" array of strings. The client should have the user
pick between these choices when that QuickFix is selected,
and pass that to the `quickfix` command as `arg`.

If the client does not (or cannot) do it there, the result of
the `quickfix` will be a non-empty list of those same
choices again, which must be presented, and the `quickfix`
command must be reissued as described above.