Experimental PromptList as UI for quickfixes with args
Refs #3

Resolving ambiguous imports works! We have a nice prompt mode
that should be pretty reusable. With some tweaks, it could be used
to handle :Implement, etc. as well.

Before merging, we should see if we can figure out how to sort the
classname choices before returning them (see the TODO in
ImportsQuickFixDescriptor). Trying to resolve List, for example,
has a bunch of rubbish before `java.util.List`---which is at the END.

There are some lingering UX issues that we should improve, like
how attempting FixProblem at a spot without problems opens an empty
window instead of just warning, and how you must be right on top
of a problem to fix it instead of just finding the closest on the line.