WIP Fix handling of "blocking" import disambiguation prompts
Refs #3

NB: OptimizeImportsTest and throwing a weird error about
the PsiFile and Document not being consistent right now

We had to save the old Problems so we could locate the ID, and
then use that to find an equivalent fix in the updated document.

This did reveal a problem with optimizing multiple *non*-ambiguous
imports, however. Theoretically, that should be handled in a
separate branch, but it may be possible to use a similar fix for
the ambiguous case with the non-ambiguous case....