Attempt to auto-fix import problems during completion
Refs #13

Specifically, if there's a single import problem, we'll attempt
to resolve it in-place and signal the client to reload, providing
the updated offset for the cursor. This provides the feel of
auto-import in some use cases, but does not improve the normal
"auto-complete classname" case.

When we attempt to complete classnames, the current implementation
will have IntelliJ return every possible classname, which is just
unreasonable. I suspect part of it is because we assumed IJ would
do the prefix detection itself, but it doesn't. However, I think
we can fairly easily pass the `a:base` param from Vim as "prefix"
to IntelliJ....