Use existing Document instances instead of forcing VimDocument
Refs #10

This is another attempt to address #10. This one feels 90% of the
way there---changes are synchronized to IJ successfully, and IJ
seems to be able to write to the file without problem. However,
all the IntelliVim commands for these open files (presumably only
those) seem to take two tries to work. Saving the file to get
problem updates, for example, will initially return some bogus
markers (not aligned with existing ones), then the second save
will provide the correct ones.

Hopefully this is as simple as ensuring the caches/indexes are updated
eagerly for such files....