Use invokeLater to allow async file updates before commands run
Refs #10

This seems to be exactly what we need, but two of the tests are
failing with the "reuse document" code. There doesn't *seem* to
be any problems in real use, but I'm hesitant to disable it just
for unit tests. More investigation is warranted to see why the
unit tests have a Document to reuse, and why that's a problem