Refactor all the obvious uses of VimEditor
Not sure if it's related to the changes here, but I discovered
an issue where Problems.collectFrom hangs.

Repro:
 - Open a file
 - Save it (trigger the command)
 - Make some trivial change causing an error and save
 - Undo and save

The last "save" there will timeout, and the call will never finish.
Will need to hop back on master and test if this repros there as well.
I can't imagine why these changes would introduce the bug (indeed,
using regular VimEditor in the method still hangs), but....