Use lookups from the completion handler's progress indicator
Refs #32

For some reason they live there, but not later. This seems to work
okay, but we have to use hacks to not crash when there's an error
(it tries to show an error hint, but there's no UI to speak of,
so... NPE).

This still doesn't seem as reliable as the old method...